+++ metadata.rb
@@ -4,7 +4,7 @@ maintainer_email 'support@atalanta-systems.com'
 license          'Apache 2.0'
 description      'Installs/Configures base'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
-version          '0.1.1'
-%w{ selinux chef-client simple_iptables }.each do |cb|
+version          '0.2.0'
+%w{ selinux }.each do |cb|
   depends cb
 end